SPEAr13xx: Append VA_ to all logical address macros
Current convention followed by SPEAr in Linux is:
- Don't append anything before macros for physical address
- Append VA_ before macros for logical address

This wasn't followed at all in all misc_regs.h for SPEAr13xx. This is corrected
and VA_ is appended to all logical address. Also, VA_ is removed from few
address which were physical (that was an mistake).

Reported-by: Vincenzo Frascino <vincenzo.frascino@st.com>

+++ arch/arm/mach-spear13xx/spear1310.c
@@ -646,7 +646,7 @@ struct pmx_dev spear1310_pmx_ras_mii_txclk = {
 /* pad multiplexing for pcie0 device */
 static struct pmx_mux_reg pmx_pcie0_mux[] = {
 	{
-		.address = VA_SPEAR1310_PCIE_SATA_CFG,
+		.address = (u32)VA_SPEAR1310_PCIE_SATA_CFG,
 		.mask = SPEAR1310_PCIE_CFG_VAL(0),
 		.value = SPEAR1310_PCIE_CFG_VAL(0),
 	},
@@ -668,7 +668,7 @@ struct pmx_dev spear1310_pmx_pcie0 = {
 /* pad multiplexing for pcie1 device */
 static struct pmx_mux_reg pmx_pcie1_mux[] = {
 	{
-		.address = VA_SPEAR1310_PCIE_SATA_CFG,
+		.address = (u32)VA_SPEAR1310_PCIE_SATA_CFG,
 		.mask = SPEAR1310_PCIE_CFG_VAL(1),
 		.value = SPEAR1310_PCIE_CFG_VAL(1),
 	},
@@ -690,7 +690,7 @@ struct pmx_dev spear1310_pmx_pcie1 = {
 /* pad multiplexing for pcie2 device */
 static struct pmx_mux_reg pmx_pcie2_mux[] = {
 	{
-		.address = VA_SPEAR1310_PCIE_SATA_CFG,
+		.address = (u32)VA_SPEAR1310_PCIE_SATA_CFG,
 		.mask = SPEAR1310_PCIE_CFG_VAL(2),
 		.value = SPEAR1310_PCIE_CFG_VAL(2),
 	},
@@ -712,7 +712,7 @@ struct pmx_dev spear1310_pmx_pcie2 = {
 /* pad multiplexing for sata0 device */
 static struct pmx_mux_reg pmx_sata0_mux[] = {
 	{
-		.address = VA_SPEAR1310_PCIE_SATA_CFG,
+		.address = (u32)VA_SPEAR1310_PCIE_SATA_CFG,
 		.mask = SPEAR1310_SATA_CFG_VAL(0),
 		.value = SPEAR1310_SATA_CFG_VAL(0),
 	},
@@ -734,7 +734,7 @@ struct pmx_dev spear1310_pmx_sata0 = {
 /* pad multiplexing for sata1 device */
 static struct pmx_mux_reg pmx_sata1_mux[] = {
 	{
-		.address = VA_SPEAR1310_PCIE_SATA_CFG,
+		.address = (u32)VA_SPEAR1310_PCIE_SATA_CFG,
 		.mask = SPEAR1310_SATA_CFG_VAL(1),
 		.value = SPEAR1310_SATA_CFG_VAL(1),
 	},
@@ -756,7 +756,7 @@ struct pmx_dev spear1310_pmx_sata1 = {
 /* pad multiplexing for sata2 device */
 static struct pmx_mux_reg pmx_sata2_mux[] = {
 	{
-		.address = VA_SPEAR1310_PCIE_SATA_CFG,
+		.address = (u32)VA_SPEAR1310_PCIE_SATA_CFG,
 		.mask = SPEAR1310_SATA_CFG_VAL(2),
 		.value = SPEAR1310_SATA_CFG_VAL(2),
 	},
